branch rjMCMC , fixed transition rates, and maximum speciation
I added a reversible jump to track the probability of speciation on target branches. Also considered a fixed transition rates estimated previously with HiSSE. Finally, I added a minimum number of speciation events expected in the whole tree to account for unobserved speciation, since we have a very incomplete sampling.