Correct Simulation and new speciationconstraint
I corrected the traderpros_sim.py so now it creates the correct species matrix file for using as inference.

I also changed the part of the algorithm for the constrained search of speciation events. I removed the species assignment ctmc node for evolution of speciation and added a distance matrix between representatives of each assigned species. It seems faster and there is no need to specify the maximum number of species in the dataset.

Also added some simulated example data. Readme file still needs to be adjusted as well as the RevBayes Tutorial.